+++ modules/features/profile/src/main/java/au/com/shiftyjelly/pocketcasts/profile/AccountDetailsFragment.kt
@@ -29,6 +29,7 @@ import au.com.shiftyjelly.pocketcasts.compose.AppTheme
 import au.com.shiftyjelly.pocketcasts.models.to.SignInState
 import au.com.shiftyjelly.pocketcasts.models.to.SubscriptionStatus
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
+import au.com.shiftyjelly.pocketcasts.profile.champion.PocketCastsChampionBottomSheetDialog
 import au.com.shiftyjelly.pocketcasts.profile.databinding.FragmentAccountDetailsBinding
 import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackManager
 import au.com.shiftyjelly.pocketcasts.repositories.playback.UpNextQueue
@@ -63,7 +64,7 @@ import au.com.shiftyjelly.pocketcasts.ui.R as UR
 import au.com.shiftyjelly.pocketcasts.views.R as VR
 
 @AndroidEntryPoint
-class AccountDetailsFragment : BaseFragment() {
+class AccountDetailsFragment : BaseFragment(), OnUserViewClickListener {
     companion object {
         fun newInstance(): AccountDetailsFragment {
             return AccountDetailsFragment()
@@ -128,6 +129,10 @@ class AccountDetailsFragment : BaseFragment() {
                     Gravatar.refreshGravatarTimestamp()
                     context?.startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(Gravatar.getGravatarChangeAvatarUrl(signInState.email))))
                 }
+            }
+
+            if (signInState.isPocketCastsChampion) {
+                binding.userView.setOnUserViewClick(this)
             }
         }
 
@@ -233,6 +238,10 @@ class AccountDetailsFragment : BaseFragment() {
             analyticsTracker.track(AnalyticsEvent.ACCOUNT_DETAILS_SHOW_TOS)
             context?.startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(Settings.INFO_TOS_URL)))
         }
+    }
+
+    override fun onPocketCastsChampionClick() {
+        PocketCastsChampionBottomSheetDialog().show(childFragmentManager, "pocket_casts_champion_dialog")
     }
 
     private fun signOut() {