+++ modules/services/analytics/src/main/java/au/com/shiftyjelly/pocketcasts/analytics/AnalyticsEvent.kt
@@ -636,4 +636,8 @@ enum class AnalyticsEvent(val key: String) {
     CLIP_SCREEN_PLAY_TAPPED("clip_screen_play_tapped"),
     CLIP_SCREEN_PAUSE_TAPPED("clip_screen_pause_tapped"),
     CLIP_SCREEN_LINK_SHARED("clip_screen_link_shared"),
+
+    /* Pocket Casts Champion */
+    POCKET_CASTS_CHAMPION_DIALOG_SHOWN("pocket_casts_champion_dialog_shown"),
+    POCKET_CASTS_CHAMPION_DIALOG_RATE_BUTTON_TAPPED("pocket_casts_champion_dialog_rate_button_tapped"),
 }